rework coeffs + modifs data
Refactor majeur
Ajout changements a la data des L1, L2 pas faite encore, 1/4 fait
Reste a réecrire les pages pour la L2
