+++ styles/global.css
@@ -59,6 +59,7 @@ body {
     line-height: 1.5em;
     text-justify: inter-word;
     text-align: justify;
+    margin-bottom: 0;
 }
 
 
@@ -239,6 +240,18 @@ main h2 a:hover {
 
 
 /* ==== [ Footer ] ==== */
+
+footer {
+    width: 100%;
+    left: 0;
+    bottom: 0;
+    position: sticky;
+    padding: 42px 0 8px 0;
+    background: linear-gradient(
+        rgba(from var(--main) r g b / 0.1),
+        var(--main) 65%
+        );
+}
 
 footer a {
     color: var(--main100);
